example/scim/requirements.txt
Jinja2
lxml
Mako
PyInstaller
python-dateutil
pyzmq
requests
